increase security by changing getSession to getUser at the backend and updated the middleware to authenticate requests and remove dotenv

+++ app/api/learning-tools/pdf/[pdfId]/delete/route.js
@@ -5,6 +5,8 @@ import { supabase } from "@/lib/supabaseClient";
 
 export async function DELETE(req, { params }) {
     try {
+        const userId = req.headers.get("x-verified-user-id");
+
         await connectToDB();
 
         const { pineconeIndex } = await getVectorStore();
@@ -18,11 +20,12 @@ export async function DELETE(req, { params }) {
         const formattedFileName = fileName.split("%20").join(" ");
 
         //delete in MongoDB
-        await PDF.findOneAndDelete({ _id: pdfId });
+        await PDF.findOneAndDelete({ _id: pdfId, userId });
 
         //delete chunks in Pinecone
         await pineconeIndex.namespace("__default__").deleteMany({
             pdfId,
+            userId,
         });
 
         //delete pdf in Supabase
@@ -36,7 +39,7 @@ export async function DELETE(req, { params }) {
             {
                 message: "Successfully deleted PDF and its components.",
             },
-            { status: 203 }
+            { status: 203 },
         );
     } catch (error) {
         console.error(error);
